Fix transition between non-flippable and flippable
Before this patch it was not possible for a window to be moved to
the origin and have it begin page flipping. Because the size didn't
change, the DRI2 buffers would never get re-allocated and the buffers
would remain without a DRM fb and therefore unable to be flipped to.

This change forces DRI2 buffers to be re-allocated whenever a
drawable transitions between being flippable and non-flippable, and
vice versa. This both allows previously non-flippable windows to
start flipping, and saves possibly scarce scanout-capable memory
when a flippable window becomes non-flippable again.

As a somewhat unrelated change, this patch also removes the
possibility of exchanging pixmaps to effect a swap. This is because
it crashed whenever we hit this case, and we don't understand enough
what is meant by exchange to be able to implement it properly.

(cherry picked from commit 4a0c0aadb21782d81bdfb4a287b72b4c473c1e5c
 from git.linaro.org)

BUG=None
TEST=None

Change-Id: I22777525cc9085f510fb0924e7981179a7cdb1ce
Signed-off-by: David Garbett <David.Garbett@arm.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/26981
Reviewed-by: Anush Elangovan <anush@google.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Ray Smith <Raymond.Smith@arm.com>
Reviewed-by: Akshu Agrawal <akshu@chromium.org>