armsoc: HACK: Avoid blitting and always flip
Blitting on the exynos5 seems to show flickering on the screen.
For some reason we seem to be failing the DRI2CanFlip() check
on suspend/resume. For now force the flip path.
This is also expected to fix one more case of the fimd irq

BUG=chrome-os-partner:11708
TEST=build armsoc and loop through 50 suspend/resume cycles

Change-Id: I154b82e73d4ee9c5b9f56a86fe3b28029825cd07
Signed-off-by: Anush Elangovan <anush@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28521
Reviewed-by: Pawel Osciak <posciak@google.com>
Commit-Ready: Anush Elangovan <anush@google.com>
Tested-by: Anush Elangovan <anush@google.com>
Reviewed-by: Mandeep Singh Baines <msb@chromium.org>