armsoc: Force page flip event even if crtc disabed
This patch will ensure that the last schedule swap is acknowledged
by the driver by completing the page flip ioctl and sending the
event for swap complete.
This fixes the issue when the screen goes "dim" and X had issued a swap
call A->B. Meanwhile before the swap was completed the crtc was disabled
and a corresponding check for this in the armsoc ddx could not complete the
page flip.
This resulted in crash when X tried to swap again from the same
buffers A->B instead of B->A.

BUG=chrome-os-partner:12083
TEST=tested with screen idle->dim->resume

Change-Id: I724420a95a85500b7bd7f9258a59bad43493c55d
Signed-off-by: Akshu Agarwal <akshu@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29074
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Anush Elangovan <anush@google.com>
Tested-by: Anush Elangovan <anush@google.com>
Commit-Ready: Abhinav Kochhar <abhinav@chromium.org>