exynos: Initialize scanout before enabling it
Be sure to initialize the contents of the new OMAP scanout resource with
the existing contents of the framebuffer prior to enabling it as our
active scanout memory.  The memory is otherwise arbitrary texture memory
until this is done so it could cause a frame of corruption when X was
first starting.

BUG=chrome-os-partner:11522
TEST=Manually tested turning X on and off with solid colours written in
to the framebuffer on Snow to verify that there was no visible
corruption.

Change-Id: I8a9c7fb9726f37392abd962238a385dc66894808
Reviewed-on: https://gerrit.chromium.org/gerrit/30463
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Ready: Jeff Disher <disher@chromium.org>
Tested-by: Jeff Disher <disher@chromium.org>