drmmode_display: comment out gamma settting
Exynos does not have any gamma setting support so don't try to
set gamma. This fixes the following error:

[     5.770] (EE) ARMSOC(0): failed to set gamma: Invalid argument

gamma is not an optional drm property so no way to really probe
for it. So only option is to comment out the code.

BUG=chrome-os-partner:13087
TEST=Verify that error message no longer appears.

Change-Id: I705601e95d1a52f60a01ed5464ecfd4f98545997
Reviewed-on: https://gerrit.chromium.org/gerrit/31763
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Ready: Mandeep Singh Baines <msb@chromium.org>
Tested-by: Mandeep Singh Baines <msb@chromium.org>