Make OMAPDRI2BufferRec keep less state
* Remove refcnt from OMAPDRI2BufferRec
Remove Yet Another Layer Of Refcounting by removing the refcount in
OMAPDRI2BufferRec.  Use Pixmap refcounts instead, and pass PixmapPtrs to
OMAPDRI2SwapComplete accordingly.

* Use ReuseBufferNotify to sync BO name with Xorg

We do some BO shuffling (especially on the flip path), so we'll need to
make sure that Xorg's accounting of the BO (a cross-process BO flink) is
pointing to the BO that we're actually holding (in the process as a BO
handle).

Do this in the ReuseBufferNotify callback so as to do this reliably
every time before Xorg is about to hand a buffer back to the
application.  This allows us to avoid having to touch the DRI2Buffer
internally every time we switch around BOs.

Also: dead code removal.

TEST=local build, run on lucas
BUG=chromium-os:151712
BUG=chrome-os-partner:14196

Change-Id: I4c535d36dbd9ecf2f01dff2d0b8981a0c9d0d15c
Reviewed-on: https://gerrit.chromium.org/gerrit/34155
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Ready: John Sheu <sheu@chromium.org>
Tested-by: John Sheu <sheu@chromium.org>