Export the real front buffer in PrepareAccess
When someone uses a 2D function on one of our per-crtc scanouts, 2D
acceleration expects the dimensions to be that of the full screen
buffer. This leads to crashes since the dimensions don't always
match. This change ensures that we expose the global scanout when
PrepareAccess is called on per-crtc scanouts.

BUG=chrome-os-partner:15427
TEST=try to take a screenshot multiple times with ctrl-F5 and
TEST=an HDMI screen plugged in, X doesn't crash.

Change-Id: I8bbe3237ec992c35a4a2416ab8fc26aba5c17264
Reviewed-on: https://gerrit.chromium.org/gerrit/36156
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Ready: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>