Correct OMAPDRI2GetMSC to query the correct display
Currently the function will always ask the kernel for the information about the
first display device, which is normally the internal display, regardless of
where the drawable is displayed. This change packs the crtc value for drawable
into the request, so the the uts and msc values returned are the ones associated
with the display that it is occuring on.

This change is needed to make my egl extension prototype
(https://gerrit-int.chromium.org/30212) behave correctly when there are multiple
monitors.

BUG=chrome-os-partner:16624
TEST=Manual. I have tested this locally, but this margin it too narrow to
     contain a proper description.

Change-Id: I223b511feaf3e0fec6111ab924fec1da7d43ffb1
Reviewed-on: https://gerrit.chromium.org/gerrit/39607
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>