Don't modify virtualX/virtualY if resize() hook fails
Excerpt from resize() hook documentation:

The driver is responsible for updating scrn->virtualX and scrn->virtualY.
If the requested size cannot be set, the driver should leave those values
alone and return FALSE.

Also refactor a bit to simplify the code.  This causes displayWidth to
get recalculated even if it doesn't change, but so be it.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=none
TEST=sanity check ui on device

Change-Id: I9bac7e5a1c0efc48772c3c930f83257f8fa6c31d
Reviewed-on: https://chromium-review.googlesource.com/190693
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>