DRI2GetMSC: Send a monotonic ust when the drawable is not full screen
According to the opengl glx_sync_control spec, the Unadjusted System Time
(or UST) is a 64-bit monotonically increasing counter that is available
throughout the system:
http://www.opengl.org/registry/specs/OML/glx_sync_control.txt

For the case where the Drawable cannot be mapped to a crtc (e.g., not full
screen, or there is no crtc), other drivers cannot just return FALSE here
as that triggers a BadDrawable error to be sent, and as is often the case
mishandled, to the client. This results in a certain compositor terminating, for example.

So, to be consistent with other drivers, return TRUE and set ust to a
monotonic system timestamp which in theory should also be monotonic with
the previous and subsequent vblank times.

Adopted from xf86-video-intel commit:
f09aa788d79d36688bcfdd3b49b92367590c5f16

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>
Suggested-by: Chris Wilson <chris@chris-wilson.co.uk>

BUG=chromium:223001
TEST=manually on daisy:
 1) attach external monitor (DP)
 2) start youtube video on external monitor
 3) uplug external monitor
 4) No regressions, and no error messages in /var/log/Xorg.0.log like:
    (EE) ARMSOC(0): ERROR: No crtc associated with given drawable
TEST=run_remote_tests --board=daisy graphics_SyncControlTest

Change-Id: I6a7dc70c79d46dce29ff3d7c5154916f849e0b96
Reviewed-on: https://gerrit.chromium.org/gerrit/47713
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>