Set resized flag when discarding scanouts
This patch sets the has_resized flag when we discard active scanout
buffers. The reason we need to do this is because the unreference here
removes the fb from the crtc in the kernel. Subsequent flips to that
crtc will fail b/c the front buffer is non-existant. Setting the
has_resized flag will ensure that we transition to blit mode before
trying to flip on that crtc (which will setcrtc to a valid fb).

BUG=None
TEST=Idle on/off works for HDMI on 3.8 kernel

Change-Id: I295f5b2cdcc1b923ff86c6f01e9eff2477032b42
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/56603