Fix leak in OMAPProbe if xf86AllocateScreen() fails
Fixes the leak of devSections and fd.

Note: The previous version of OMAPProbe returned FALSE if *any*
xf86AllocateScreen() fails. This version returns TRUE if at least one
instance is successfully claimed.  It should be ok to only allocate
and claim some of the screens that were specified by xorg.conf "Device"
sections.

TODO: Opening fd with OMAPOpenDRM() and closing it with drmClose() is a
mismatch.  Better would be to:
 * open with one of drmOpen*()
 * open with open() and close with close()
 * introduce a new OMAPCloseDRM()
But, any of these would require a separate patch.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=none
TEST=builds clean; sanity check ui

Change-Id: Iceb2a5349fa87f57f83719f01e4094ec31ed4b6d
Reviewed-on: https://chromium-review.googlesource.com/65848
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>