drmmode: store just the plane_id, not drmModePlane
To access the cursor's plane, we only need its DRM ID.  There is no need
to carry around the full drmModePlane struct.  In fact, there is no
reason to fetch the drmModePlane structure at all, since all we do is
save its plane_id, which we already knew to fetch it in the first place.

This frees a tiny amount of memory sooner and makes the cleanup paths a
little simpler.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:260197
TEST=manual: inspect /var/log/Xorg.0.log
TEST=driver loads
TEST=cursor loaded, shown, hidden and updated appropriately

Change-Id: I720d4a58612d2e3240a74c5d0f3e3ef8fc67e9fb
Reviewed-on: https://chromium-review.googlesource.com/168186
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>