armsoc: Support multiple acquires of the same buffer
Allow multiple clients to acquire buffer for read,
also if buffer is acquired for write, allow others to use it
for reads so a blit in that uses the same buffer for source
and destination can succeed. It is not possible to allow
acquire for write if a shared (read) acquire is already in effect.
It would require releasing a lock in kernel and reacquiring it for
exclusive access while someone is possibly accessing the buffer.
So this condition will trigger an error and acquire failure.
Fortunately Xserver does not do that.
Also, reorder acquires for internal blit to acquire destination for
write first.

Change-Id: Ia6c69d27a2029cce96e3d8f47f3a6cb73a10f038
Signed-off-by: Dominik Behr <dbehr@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/170074
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>