armsoc: wait for all flips to finish before switching to blit mode
When switching to blit mode we copy per-CRTC scanout buffer back to main
buffer. If there are any flips in flight, we copy stale contents so screen
capture gets stale data and whats worse, when swicthing back to flip mode
we restore this stale data back to last buffer in the flip chain and it
will look like we are going back in time. Waiting for all flips to complete
fixes it because it makes us copy from the newest buffer.

BUG=chromium:233233
TEST=go to https://www.corp.google.com/~sheu/webrtc_screencast/

Change-Id: I3cf322c9462bc97516059e36f8095a4cac538295
Signed-off-by: Dominik Behr <dbehr@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/170095
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>