Add fb ids only when needed
Adding fb id has been moved to omap_bo_get_fb. So fb id is now created only
when it is needed. Explicit calls to omap_bo_add_fb have been removed. Also
omap_bo_get_fb is only called when necessary. This also fixes the bug with
2nd display losing ability to flip after closing and opening the lid because
fb id for the drawable was only added at creation time when it did not
match any scanout position and size. Now it is added when it is needed and
the decision whether to flip or not is made only in one place in SwapBuffers.

BUG=chromium:315957
TEST=move primary display to 2nd, run webgl aquarium, close and open lid
repeatedly, observe frame rate

Change-Id: I393dbfd7b0b2355d5310a36f080e4b90a2038764
Signed-off-by: Dominik Behr <dbehr@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/176284
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>