Move logging/debug message macros into their own header file
It would be convenient to let the dumb buffer layer share the same logging
macros as the rest of the driver so that its messages will have the same
header.

However, we don't want omap_dumb to depend on all the other things defined
in omap_driver.h.  So, split out just the msg definitions to its own
header file.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:353437
TEST=sanity check UI w/ and w/out Debug uncommented in
  /etc/X11/xorg.conf.d/exynos.conf

Change-Id: I9bf6b02bbc752edcafef41aebc908e545dd73ad2
Reviewed-on: https://chromium-review.googlesource.com/190688
Reviewed-by: John Sheu <sheu@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Dominik Behr <dbehr@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>