omap_dumb: use armsoc ERROR/DEBUG_MSG() macros
First we plumb ScrnInfoPtr down into the omap_dumb layer, and save it in
the omap_dev struct.  Then, we can use it to print error/debug messages
with a proper xf86DrvMsg() header.

While we are here, we fix up the error messages a bit:
 * use "%s" % (strerror(errno)) to print errno messages
 * use "[BO:%u]" % (bo->handle) for identifying buffer objects
 * always use a single line for string literals (to improve grep-ability)

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:353437
TEST=sanity check UI w/ and w/out Debug uncommented in
  /etc/X11/xorg.conf.d/exynos.conf

Change-Id: I81e14f3d0333297096191d8a5ae0a1e44742056a
Reviewed-on: https://chromium-review.googlesource.com/190689
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>