omap_dumb: add some bo/fb lifecycle logging
BO and FB lifecycle events (creation, mapping, deletion, etc.) are
relatively rare.  It makes tracing and debugging easier if we have nice
informative logging.

Actually, this shows that there are a lot of small bos created and
destroyed at driver initialization.  These buffer allocations probably
don't need to be going all the way down to the kernel, but that is
for another patch to fix.

We use the armsoc DEBUG_MSG() macro for logging BO lifecycle events.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:272670
TEST=sanity check UI, and enjoy the new BO & FB lifecycle event logging in
   /var/log/Xorg.0.log
   In normal usage, there should be minimal log growth - only at X server
   start, X server shutdown, and monitory hotplug events.

Change-Id: I209a778eda00a65204411f9b54a66218da9fa92d
Reviewed-on: https://chromium-review.googlesource.com/65852
Reviewed-by: John Sheu <sheu@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>