exa: Update logic for choosing root pixmap's backing bo in prepare access
The Pixmap for non-redirected windows is the root Pixmap.  Traditionally,
the root Pixmap is backed by a BO with its same dimensions.  Thus, its
->devPrivate.ptr pointed to a buffer that had the same dimensions and
pitch as the root Pixmap itself.

When creating a set of DRI2 buffers for a DRI2 drawable, ARMSOC creates
a new Pixmap for its back buffer, with its own BO, but re-uses the root
Pixmap for its front buffer.

When ARMSOC receives a DRI2 SwapBuffers request for an un-clipped window
whose dimensions match one of the CRTCs, it switches to "flip" mode,
where it programs the CRTCs to scan out directly from the "per-CRTC" BOs.
When the flip completes, it "exchanges" the front and back Pixmaps'
backing BOs.  Thus, the "backing BO" for the root Pixmap actually changes
on each flip when in flip mode.

ARMSOC provides 2D access to Pixmaps via EXA.  Before actually accessing
a Pixmap, the EXA layer calls ARMSOC's OMAPPrepareAccess().
OMAPPrepareAccess() sets the Pixmap's ->devPrivate.ptr to point to a
buffer that (a) has up-to-date pixel data for reading and/or is accessible
for writing, (b) has the same size as the Pixmap itself, and (c) won't be
written to by others during this access.

For the root Pixmap, (a) and (b) must be handled carefully, since as
mentioned above, the relationship between the root Pixmap and its backing
BO is complicated during flip mode.  In particular:
 * For writes, we never give access to a per-CRTC BO, all writes must be
   to the root BO, and therefore we switch to blit mode.
 * For reads, since 2D may access any pixels of the root Pixmap we must
   ensure that the BO we give back has the the same dimensions (and pitch).
   * When using just a single CRTC, its per-CRTC BO will have the same
     dimensions as the root Pixmap and therefore, as an optimization, we
     can just provide 2D with direct read access to it.
   * When using multiple CRTCs, however, the current per-CRTC BO will have
     different dimensions than the root Pixmap, so we give back the root
     Pixmap after first updating its contents from all per-CRTC BOs.

In all cases, we grab a lock on the per-CRTC BO to hold off any updates
until this read access has completed.

The big change in this patch is how we determine whether the Pixmap for
whom access is being prepared is the root Pixmap.  Previously, we would
only treat it as the root Pixmap if its backing BO was the same as the BO
of one of the per-CRTC scanouts.  This is almost always correct, but
fails in one special case:
  If we have multiple CRTCs, and we just switched modes, but the new mode
does not change the dimensions of the root window, and the last Swap
completed was for a window on the CRTC whose mode has just changed, then
the root Pixmap's backing BO will be the old per-CRTC BO.  However, since
we just changed modes, that old per-CRTC BO will no longer be the BO for
per-CRTC scanout, having been replaced by a new BO with the correct size
for the new mode.  Thus we won't find the BO currently backing the root
Pixmap in the per-CRTC scanout list, and erroneously give it back.  Since
this per-CRTC BO does not have the root Pixmap's dimensions, the X server
will segfault when trying to access pixels outside of the range of that
old CRTC.
  As an example, if we have two CRTCs 1366x768 and 740x480, they are
arranged on a 1366x1308 root window like this:

       0 ....    740 ....  1366
      ++-----------+-----------+
    0 |+----------------------+|
 .... ||       1366x768       ||
  767 |+----------------------+|
  768 |/**********************\|
 .... |I                      I|
  827 |\**********************/|
  828 |+-----------+/*********\|
 .... || 740x480   |I         I|
 1307 |+-----------+\*********/|
 1308 +------------------------+

If we then change the bottom CRTC to 640x480, the root Pixmap dimensions
don't change - it is still 1366x1308.  Thus, this mode change does not
cause a root window resize, so we do not modify its Pixmap header, nor
allocate a new root BO.  Thus, immediately after the mode change, the
root Pixmap's backing BO is still whatever it was before. If the last swap
completed was a swap of the 640x480 CRTC, then the root Pixmap's backing
BO will still be an old 640x480 BO.

When we did the mode change on the lower CRTC, however, we did allocate
a new BO for its per-CRTC scanout.  That new BO is now in pOMAP->scanouts[]
array.

Now, if at this moment an X Client requests to draw on the root window,
we will check the root Pixmap's backing BO in the per-CRTC scanout list,
but not find it.  And therefore incorrectly give back the old 640x480 BO.
If the X Client requested access to pixels outside of this (too-small)
buffer, the X server segfaults (ie in the corresponding pixman drawing
call).

This is actually what happens in Chrome OS:
   Chrome requests via a new CRTC configuration in XRandR, the X server
computes a new root window to contain this CRTC configuration, modifies
the root Pixmap if needed, and then instructs ARMSOC to reprogram the
per-CRTC modes.  Chrome then clears the area between the new CRTCs by
calling XFillRectangle() on the root window.  If we are unlucky and the
last DRI2SwapBuffers processed was for the external monitor, then the root
Pixmap will be backed by its old per-crtc bo, which is no longer
one of the BOs in pOMAP->scanouts[].  Thus, we end up accessing the much
smaller per-CRTC BO instead of the root BO which causes a segfault.


Instead of deducing that the Pixmap is the root Pixmap by comparing its
backing BO to the BOs in per-CRTC scanout list, we take a more direct
approach - just compare the Pixmap itself with the current root Pixmap of
the screen.

This patch also fixes a related bug.  Previously, when providing write
access to the root Pixmap, we were only invalidating the per-CRTC scanout
corresponding to the root Pixmap's current backing BO.  However, 2D write
access to the root Pixmap could write anywhere on the root Pixmap, not
just to the region corresponding to the current per-CRTC BO.  Thus, we
must invalidate all per-CRTC BOs.  This patch implements this by just
invalidating all per-CRTC scanouts when switching to blit mode.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:27725
TEST=See the steps in crosbug.com/p/27725#c19

Change-Id: I0ef17ce1eda3a66a84178d7d5eef087c837e6438
Reviewed-on: https://chromium-review.googlesource.com/193741
Tested-by: shirish s <s.shirish@samsung.com>
Reviewed-by: shirish s <s.shirish@samsung.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Abhinav Kochhar <abhinav.k@samsung.com>
Commit-Queue: shirish s <s.shirish@samsung.com>