drmmode: do not clear newly allocated bos used for scanouts
The kernel always clears memory for when allocating a buffer for a GEM
object.  There is no reason to clear it manually in user space.

This saves a few milliseconds when the X server is starting
(~4 for a single 1366x768 monitor, about ~24 when using a
1366x768 & 1920x1080).

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:26503
TEST=manually verified that new gem buffers are in fact filled with 0s.
   => No corrupted framebuffers at HDMI plug/unplug

Change-Id: Ib716bd611e023c9fc21ef883346882cda4fae78e
Reviewed-on: https://chromium-review.googlesource.com/191063
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>