drmmode: don't fetch drmModeCrtc
Armosc doesn't use "current crtc state".  All it ever needs is the
crtc_id - which it already had to know to call drmModeGetCrtc in the first
place.

While we are at it, use drmmode_crtc_id() everywhere to fetch the crtc_id,
and mark that function static since it is only used in this file.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:370411
TEST=sanity check ui

Change-Id: Ifef42da0a7e9ea64355b490451a3b3e7ed2716e7
Reviewed-on: https://chromium-review.googlesource.com/198368
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>