Only fetch request property in .get_property()
This patch changes the behavior of the .get_property() callback to only fetch
the request property value from kernel. The old behavior was to update the
entire connector when any property was requested. This has the side-effect of
also fetching the EDID, which can be time-consuming.

The downside to this method is that a property which has been newly attached to
the connector since the last getConnector will no longer be returned. This is
reasonable behavior since properties are added at initialization time. We could
work around this by fetching the connector state if the property is not found,
but I don't think it's worthwhile.

BUG=chromium:374465
TEST=Tested on peach, xrandr properly updates property values

Change-Id: I284254103aa5bf14238f6c44c82faeb120e0113e
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/208332
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>